Fixed Account creation and login system.

+++ app.py
@@ -1,4 +1,4 @@
-from flask import Flask, render_template, flash, request, redirect, url_for, sessions
+from flask import Flask, render_template, flash, request, redirect, url_for, session
 from flask_wtf import FlaskForm
 from wtforms import StringField, SubmitField, PasswordField, BooleanField, ValidationError, SelectField, IntegerField, \
     validators
@@ -10,14 +10,13 @@ import yaml
 import mysql.connector
 from mysql.connector import errorcode
 
-# !!!!! Clean up, remove all tutorial junk into separate file, doesnt matter it runs or not
-
-
 # !!!! Remove instructor table
 # !!!! Add location to lesson table
 
 # !!! Need Belt id not name in account object
 # !!! class to add new account must account for relations
+
+# !! Passwords dont match
 
 # !! CREATE TABLE CREATION STATEMENTS
 # !! REMOVE CAPACITY TABLE
@@ -51,37 +50,41 @@ db = yaml.safe_load(open('db.yaml'))
 # !!!WARNING!!! DEV TOOL ONLY
 # This deletes all tables from current database in order to refresh them. This method is extremely destructive and
 # must be used with caution
-def DeleteAllTables(cursor):
-    try:
-        cursor.execute("DROP TABLE BOOKINGS")
-    except:
-        print("Table doesn't exist")
+def DeleteAllTables(cursor, key):
+    if key == db['db_wipe_key']:
+        try:
+            cursor.execute("DROP TABLE BOOKINGS")
+        except:
+            print("Table doesn't exist")
 
-    try:
-        cursor.execute("DROP TABLE LESSONS")
-    except:
-        print("Table doesn't exist")
-    try:
-        cursor.execute("DROP TABLE ACCOUNTS")
-    except:
-        print("Table doesn't exist")
+        try:
+            cursor.execute("DROP TABLE LESSONS")
+        except:
+            print("Table doesn't exist")
+        try:
+            cursor.execute("DROP TABLE ACCOUNTS")
+        except:
+            print("Table doesn't exist")
 
-    try:
-        cursor.execute("DROP TABLE USERS")
-    except:
-        print("Table doesn't exist")
+        try:
+            cursor.execute("DROP TABLE USERS")
+        except:
+            print("Table doesn't exist")
 
-    try:
-        cursor.execute("DROP TABLE BELTS")
-    except:
-        print("Table doesn't exist")
+        try:
+            cursor.execute("DROP TABLE BELTS")
+        except:
+            print("Table doesn't exist")
+    else:
+        print("\n\n" + "-" * 60 + "\n\nDatabase wipe attempted and failed due to invalid key.\n\n" + '-' * 60)
 
 
-# SQL Statments for table check and creation
+# SQL Statements for table check and creation
+
 TABLES = {'BELTS': '''CREATE TABLE BELTS(
                    belt_id int PRIMARY KEY,
                    belt_name char(50) NOT NULL,
-                   wait_time DATE);''', 'USERS': '''CREATE TABLE USERS(
+                   wait_time DATE);''', 'USERS': ''' CREATE TABLE USERS(
                    user_id int PRIMARY KEY,
                    belt_id int NOT NULL,
                    first_name char(50) NOT NULL,
@@ -111,14 +114,14 @@ TABLES = {'BELTS': '''CREATE TABLE BELTS(
                       FOREIGN KEY (lesson_id) REFERENCES Lessons(lesson_id),
                       FOREIGN KEY (user_id) REFERENCES Users(user_id));'''}
 
-
 try:  # try statement to provide user-friendly error messages if any database-related errors are raised
     mydb = mysql.connector.connect(  # configure connection to mysql database
         host=db['mysql_host'],
         user=db['mysql_user'],
         password=db['mysql_password'],
         port=db['mysql_port'],
-        database=db['mysql_db'])
+        database=db['mysql_db'],
+        auth_plugin='mysql_native_password')
 except mysql.connector.Error as err:
     if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
         print("Unable to authorise access to database")
@@ -127,7 +130,8 @@ except mysql.connector.Error as err:
     else:
         print(err)
 
-mycursor = mydb.cursor()
+# Create cursor to run commands to database
+mycursor = mydb.cursor(buffered=True)
 
 
 # CREATE THE TABLES WHICH DO NOT EXIST
@@ -178,8 +182,9 @@ def table_check(cursor):
 
 
 # Checks to see if any tables are missing
-DeleteAllTables(mycursor)
+
 table_check(mycursor)
+mydb.commit()
 
 # CREATE DATABASE TABLES IF MISSING
 
@@ -256,7 +261,7 @@ class Account:
         self._id = id
         self._email = None
         self._password_hash = None
-        self._user_id = None
+        self._user_id = id
         self._authority = None
         self._last_logged_in = None
         self._date_added = None
@@ -275,158 +280,55 @@ class Account:
 
     @property
     def email(self):
-        if self._email is None:
-            mycursor.execute('SELECT email FROM accounts WHERE id = %s', [self._id])
-            email = mycursor.fetchall()
-            email = email[0][0]
-            self._email = email
-            return email
-        else:
-            return self._email
+        return self._email
 
     @property
     def password_hash(self):
-        if self._password_hash is None:
-            mycursor.execute('SELECT password_hash FROM accounts WHERE id = %s', [self._id])
-            password = mycursor.fetchall()
-            password = password[0][0]
-            self._password_hash = password
-            return password
-        else:
-            return self._password_hash
+        return self._password_hash
 
     @property
     def user_id(self):
-        if self._user_id is None:
-            mycursor.execute('SELECT user_id FROM accounts WHERE id = %s', [self._id])
-            user_id = mycursor.fetchall()
-            user_id = user_id[0][0]
-            self._user_id = user_id
-            return user_id
-        else:
-            return self._user_id
+        return self._user_id
 
     @property
     def authority(self):
-        if self._authority is None:
-            mycursor.execute('SELECT authority FROM accounts WHERE id = %s', [self._id])
-            authority = mycursor.fetchall()
-            authority = authority[0][0]
-            self._authority = authority
-            return authority
-        else:
-            return self._authority
+        return self._authority
 
     @property
     def last_logged_in(self):
-        if self._last_logged_in is None:
-            mycursor.execute('SELECT last_logged_in FROM accounts WHERE id = %s', [self._id])
-            last_logged_in = mycursor.fetchall()
-            last_logged_in = last_logged_in[0][0]
-            self._last_logged_in = last_logged_in
-            return last_logged_in
-        else:
-            return self._last_logged_in
+        return self._last_logged_in
 
     @property
     def date_added(self):
-        if self._date_added is None:
-            mycursor.execute('SELECT date_added FROM accounts WHERE id = %s', [self._id])
-            date_added = mycursor.fetchall()
-            date_added = date_added[0][0]
-            self._date_added = date_added
-            return date_added
-        else:
-            return self._date_added
+        return self._date_added
 
     @property
     def first_name(self):
-        if self._first_name is None:
-            mycursor.execute('SELECT first_name FROM users INNER JOIN accounts ON users.user_id = accounts.user_id '
-                             'WHERE '
-                             'accounts.id = %s', [self._id])
-            first_name = mycursor.fetchall()
-            first_name = first_name[0][0]
-            self._first_name = first_name
-            return first_name
-        else:
-            return self._first_name
+        return self._first_name
 
     @property
     def last_name(self):
-        if self._last_name is None:
-            mycursor.execute('SELECT last_name FROM users INNER JOIN accounts ON users.user_id = accounts.user_id '
-                             'WHERE '
-                             'accounts.id = %s', [self._id])
-            last_name = mycursor.fetchall()
-            last_name = last_name[0][0]
-            self._last_name = last_name
-            return last_name
-        else:
-            return self._last_name
+        return self._last_name
 
     @property
     def age(self):
-        if self._age is None:
-            mycursor.execute('SELECT age FROM users INNER JOIN accounts ON users.user_id = accounts.user_id WHERE '
-                             'accounts.id = %s', [self._id])
-            age = mycursor.fetchall()
-            age = age[0][0]
-            self._age = age
-            return age
-        else:
-            return self._age
+        return self._age
 
     @property
     def belt_id(self):
-        if self._belt_id is None:
-            mycursor.execute('SELECT belt_id FROM users INNER JOIN accounts ON users.user_id = accounts.user_id '
-                             'WHERE '
-                             'accounts.id = %s', [self._id])
-            belt_id = mycursor.fetchall()
-            belt_id = belt_id[0][0]
-            self._belt_id = belt_id
-            return belt_id
-        else:
-            return self._belt_id
+        return self._belt_id
 
     @property
     def last_graded(self):
-        if self._last_graded is None:
-            mycursor.execute('SELECT last_graded FROM users INNER JOIN accounts ON users.user_id = accounts.user_id '
-                             'WHERE '
-                             'accounts.id = %s', [self._id])
-            last_graded = mycursor.fetchall()
-            last_graded = last_graded[0][0]
-            self._last_graded = last_graded
-            return last_graded
-        else:
-            return self._last_graded
+        return self._last_graded
 
     @property
     def instructor_id(self):
-        if self._instructor_id is None:
-            mycursor.execute('SELECT instructor_id FROM instructors INNER JOIN users ON instructors.user_id = '
-                             'users.user_id INNER JOIN '
-                             'accounts ON users.user_id = accounts.user_id WHERE '
-                             'accounts.id = %s', [self._id])
-            instructor_id = mycursor.fetchall()
-            instructor_id = instructor_id
-            self._instructor_id = instructor_id
-            return instructor_id
-        else:
-            return self._instructor_id
+        return self._instructor_id
 
     @property
     def location(self):
-        if self._location is None:
-            mycursor.execute('SELECT location FROM instructors WHERE instructor_id = %s', [self.instructor_id])
-            location = mycursor.fetchall()
-            location = location[0][0]
-            self._location = location
-            return location
-        else:
-            return self._location
+        return self._location
 
     # ------------- Setters -------------
 
@@ -444,7 +346,7 @@ class Account:
 
     @user_id.setter
     def user_id(self, user_id):
-        self.user_id = user_id
+        self._user_id = user_id
 
     @authority.setter
     def authority(self, authority):
@@ -486,6 +388,29 @@ class Account:
     def location(self, location):
         self._location = location
 
+def conv_id_obj(id):
+    where_parameter = (str(id), )
+    sql_fetch_all_accounts = "SELECT * FROM accounts WHERE id = %s"
+    sql_fetch_all_users = "SELECT * FROM users WHERE user_id = %s"
+    mycursor.execute(sql_fetch_all_accounts, where_parameter)
+    account_data = mycursor.fetchall()
+    mycursor.execute(sql_fetch_all_users, where_parameter)
+    user_data = mycursor.fetchall()
+
+    current = Account(id)
+
+    current.user_id = id
+    current.email = account_data[0][2]
+    current.password_hash = account_data[0][3]
+    current.authority = account_data[0][4]
+    current.date_added = account_data[0][5]
+    current.last_logged_in = account_data[0][6]
+    current.belt_id = user_data[0][1]
+    current.first_name = user_data[0][2]
+    current.last_name = user_data[0][3]
+    current.last_graded = user_data[0][4]
+
+    return current
 
 # Function to update an account object to the database
 def update_to_db(account):
@@ -532,21 +457,17 @@ def insert_account_into_db(account):
         users_last_graded = account.last_graded
 
         sql_insert_into_accounts = "INSERT INTO accounts (id, user_id ,email, password_hash, authority, " \
-                                   "last_logged_in, date_added) VALUES (%s, %s, %s, %s, %s, %s, %s)", (accounts_id,
-                                                                                                       foreign_user_id,
-                                                                                                       accounts_email,
-                                                                                                       accounts_password_hash,
-                                                                                                       accounts_authority,
-                                                                                                       accounts_last_logged_in,
-                                                                                                       accounts_date_added)
+                                   "last_logged_in, date_added) VALUES (%s, %s, %s, %s, %s, %s, %s) "
+        accountValues = (accounts_id, foreign_user_id, accounts_email, accounts_password_hash, accounts_authority,
+                         accounts_last_logged_in, accounts_date_added)
 
         sql_insert_into_users = "INSERT INTO users (user_id, first_name, last_name, age, belt_id,last_graded) VALUES " \
-                                "(%s, %s, %s, %s, %s, %s)", (
-                                    foreign_user_id, users_first_name, users_last_name, users_age, foreign_belt_id,
-                                    users_last_graded)
+                                "(%s, %s, %s, %s, %s, %s) "
+        userValues = (foreign_user_id, users_first_name, users_last_name, users_age, foreign_belt_id, users_last_graded)
 
-        mycursor.execute(sql_insert_into_accounts)
-        mycursor.execute(sql_insert_into_users)
+        mycursor.execute(sql_insert_into_users, userValues)
+        mycursor.execute(sql_insert_into_accounts, accountValues)
+
         mydb.commit()
 
     except mysql.connector.Error as error:
@@ -569,19 +490,23 @@ def index():  # put application's code here
 def login():
     form = LoginForm()
     if form.validate_on_submit():
-        account = Accounts.query.filter_by(email=form.email.data).first()
-        print(account)
-        if account:
-            # Check Hashed Password
-            if check_password_hash(account.password_hash, form.password.data):
-                login_user(account)
-                flash("Login Successful!", "success")
+        try:
+            find_account = "SELECT id FROM accounts WHERE email = %s"
+            where_condition = (form.email.data, )
+            mycursor.execute(find_account,where_condition)
+            login_id = mycursor.fetchone()
+            account_data_all = conv_id_obj(login_id[0])
 
-                return redirect(url_for('dashboard'))
+            # Check Hashed Password with inputted one
+            if check_password_hash(account_data_all.password_hash, form.password.data):
+                flash("Login Successful!", "success")
+                return render_template('test.html', current_user = account_data_all)
             else:
-                flash("Wrong Password - Try Again!", "warning")
-        else:
-            flash("Wrong Password or Username. Try Again...", category="warning")
+                flash("Wrong Password or Username. Try Again...", category="danger")
+        except:
+            print("account doesn't exist")
+            flash("Wrong Password or Username. Try Again...", category="danger")
+
     return render_template('login.html', form=form)
 
 
@@ -607,20 +532,18 @@ def dashboard():
 
 @app.route('/create/student', methods=['GET', 'POST'])
 def create_student():
-    print("hello")
     form = NewStudentForm()
     account = None
-    print("test6")
+
     if form.validate_on_submit():
-        print("test5")
         search_account_statement = "SELECT * FROM accounts WHERE email = %(email)s"
         parameter = {'email': form.email.data}
         mycursor.execute(search_account_statement, parameter)
         account = mycursor.fetchone()
-        print("test4")
+
         if account is None:
             mycursor.execute("SELECT id FROM accounts ORDER BY id DESC")
-            print("test3")
+
             next_id = mycursor.fetchone()
             try:
                 new_id = int(next_id[0]) + 1
@@ -629,7 +552,6 @@ def create_student():
                 new_id = 1
             new_account = Account(new_id)
 
-            print("test2")
             # Hash password
             hashed_pw = generate_password_hash(form.password.data, "sha256")
             # find belt in database
@@ -640,12 +562,13 @@ def create_student():
             new_account.authority = form.authority.data
             new_account.age = form.age.data
             new_account.date_added = datetime.utcnow()
-
-            print("test1")
+            new_account.password_hash = hashed_pw
+            new_account.last_logged_in = datetime.utcnow()
+            new_account.last_graded = datetime.utcnow()
             insert_account_into_db(new_account)
-            print("test19087097")
+
             flash("User Added Successfully!", 'success')
-            return redirect(url_for('account_details', id=account.id))
+            return redirect(url_for('account_details', id=new_account.id))
 
         form.first_name.data = ''
         form.last_name.data = ''
@@ -721,4 +644,3 @@ def page_not_found(e):
 # Run the App
 if __name__ == '__main__':
     app.run()
-
